Move LensWrap to widget module
