+++ druid/src/widget/mod.rs
@@ -28,6 +28,7 @@ mod identity_wrapper;
 mod image;
 mod invalidation;
 mod label;
+mod lens_wrap;
 mod list;
 mod padding;
 mod painter;
@@ -64,6 +65,7 @@ pub use env_scope::EnvScope;
 pub use flex::{CrossAxisAlignment, Flex, FlexParams, MainAxisAlignment};
 pub use identity_wrapper::IdentityWrapper;
 pub use label::{Label, LabelText, LineBreaking};
+pub use lens_wrap::LensWrap;
 pub use list::{List, ListIter};
 pub use padding::Padding;
 pub use painter::{BackgroundBrush, Painter};
